Refactor to support tracking upload
